storing new cards
trying to use fs.writeFile and fs.appendFile to save new cards that user add.
Original plan is push to the 'cards' variable in log.js, because it is where the questions come from. But, I didn't make it work. Therefore, I try to write to another .txt file, which achieve data output and save at least.